move reflectors to script_bindings
